+++ vitis/software/Baremetal/src/uz/uz_ParameterID/FrictionID/uz_PID_FrictionID.c
@@ -17,7 +17,7 @@
 #if UZ_PARAMETERID_ACTIVE > 0U
 #include "uz_PID_FrictionID.h"
 
-static size_t instances_counter_PID_FrictionID = 0;
+static uint32_t instances_counter_PID_FrictionID = 0;
 
 static uz_PID_FrictionID_t instances_PID_FrictionID[UZ_PARAMETERID_ACTIVE] = { 0 };
