+++ vitis/software/Baremetal/src/uz/uz_ParameterID/OnlineID/AutoRefCurrents/uz_PID_AutoRefCurrents.c
@@ -17,7 +17,7 @@
 #if UZ_PARAMETERID_ACTIVE > 0U
 #include "uz_PID_AutoRefCurrents.h"
 
-static size_t instances_counter_PID_AutoRefCurrents = 0;
+static uint32_t instances_counter_PID_AutoRefCurrents = 0;
 
 static uz_PID_AutoRefCurrents_t instances_PID_AutoRefCurrents[UZ_PARAMETERID_ACTIVE] = { 0 };
