+++ vitis/software/Baremetal/src/uz/uz_ParameterID/TwoMassID/uz_PID_TwoMassID.c
@@ -17,7 +17,7 @@
 #if UZ_PARAMETERID_ACTIVE > 0U
 #include "uz_PID_TwoMassID.h"
 
-static size_t instances_counter_PID_TwoMassID = 0;
+static uint32_t instances_counter_PID_TwoMassID = 0;
 
 static uz_PID_TwoMassID_t instances_PID_TwoMassID[UZ_PARAMETERID_ACTIVE] = { 0 };
